feat(post-deploy1): Make inputs in WriteProgressions wider. Organize lib/data folder to remove unneeded. Condense all initialStates in to initialStates file. Change imports to those affected files.
