+++ app/components/ExamPages/3-5_ScalesNotate.tsx
@@ -10,7 +10,7 @@ import {
   Typography,
 } from "@mui/material";
 
-import { notationInstructions } from "@/app/lib/instructions";
+import { notationInstructions } from "@/app/lib/data/instructions";
 import { FormEvent, UserDataProps } from "@/app/lib/typesAndInterfaces";
 import { useState } from "react";
 import CardFooter from "../CardFooter";
