--- app/lib/data/initialChordData.ts
@@ -1,8 +1,0 @@
-import { Chord } from "../typesAndInterfaces";
-
-export const initialChordData: Chord = {
-  keys: [],
-  duration: "w",
-  staveNotes: null,
-  userClickY: 0,
-};
