--- app/lib/data/initialNotesAndCoordinatesState.ts
@@ -1,7 +1,0 @@
-import { NotesAndCoordinatesData } from "../typesAndInterfaces";
-export const initialNotesAndCoordsState: NotesAndCoordinatesData = {
-  note: "",
-  originalNote: "",
-  yCoordinateMin: 0,
-  yCoordinateMax: 0,
-};
