--- app/lib/data/initialScaleState.ts
@@ -1,9 +1,0 @@
-import { ScaleData } from "../typesAndInterfaces";
-
-export const initialScaleState: ScaleData = {
-  keys: [""],
-  duration: "q",
-  staveNote: null,
-  userClickY: 0,
-  staveNoteAbsoluteX: 0,
-};
